refactor: restructure stock adjustments page to match sales orders split-screen layout
