+++ frontend/src/App.tsx
@@ -14,6 +14,7 @@ const ProductsPage = React.lazy(() => import('./pages/inventory/ProductsPage'))
 const CreateProductPage = React.lazy(() => import('./pages/inventory/CreateProductPage'))
 const CategoriesPage = React.lazy(() => import('./pages/inventory/CategoriesPage'))
 const StockAdjustmentsPage = React.lazy(() => import('./pages/inventory/StockAdjustmentsPage'))
+const CreateStockAdjustmentPage = React.lazy(() => import('./pages/inventory/CreateStockAdjustmentPage'))
 const SalesPage = React.lazy(() => import('./pages/sales/SalesPage'))
 const CustomersPage = React.lazy(() => import('./pages/sales/CustomersPage'))
 const OrdersPage = React.lazy(() => import('./pages/sales/OrdersPage'))
@@ -67,6 +68,7 @@ function App() {
                     <Route path="/inventory/products/:id/edit" element={<CreateProductPage />} />
                     <Route path="/inventory/categories" element={<CategoriesPage />} />
                     <Route path="/inventory/stock-adjustments" element={<StockAdjustmentsPage />} />
+                    <Route path="/inventory/stock-adjustments/create" element={<CreateStockAdjustmentPage />} />
 
                     {/* Sales Management */}
                     <Route path="/sales" element={<SalesPage />} />
